==> CODE/4_matrix_data_cleanup.ipynb <==
# Toad matrix

## Read in data and look at some features

import pandas as pd
import numpy as np

df = pd.read_csv('../DATA/2017_data_cleaned.csv')

df.columns

df.head()

len(df)

df_clean = df[~df['Adjusted_sex'].isna()]

len(df_clean)

df_clean.SVL_mm.unique()

## Create rows for each entry in the matrix
Rows contain X, Y, SVL and demog for toad1 and toad2, as well as the distance between toad1 and toad2 and diffencence in SVL.


# set to the correct df!!
dataset = df_clean

rows = []
ntoads = len(dataset)
for i, x in dataset.iterrows():
    ax, ay, asvl, asex = x[['X', 'Y', 'SVL_mm', 'Adjusted_sex']].values
    for j, y in dataset.iterrows():
        bx, by, bsvl, bsex = y[['X', 'Y', 'SVL_mm', 'Adjusted_sex']].values
        try:
            dist = np.sqrt((ax - bx)**2 + (ay -by)**2)
            diff = np.abs(asvl-bsvl)
        except TypeError as e:
            print(ax, ay, asvl, asex, asvl)
            print(bx, by, bsvl, bsex, bsvl)
            print('---')
            raise e

        row = {
            'id1': i,
            'id2': j,
            'svl1': asvl,
            'svl2': bsvl,          
            'sex1': asex,
            'sex2': bsex,
            'dist': dist,
            'diff': diff,
        }
        rows.append(row)

len(rows)

output_df = pd.DataFrame.from_dict(rows)

len(output_df)

output_df.head()

output_df.to_csv('../DATA/20220117_2017_matrix.csv')


1151*1151
